requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
asgiref==3.8.1
attrs==24.3.0
build==1.2.2.post1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==5.1.3
frozenlist==1.5.0
gunicorn==23.0.0
idna==3.10
multidict==6.1.0
oauthlib==3.2.2
packaging==24.2
pillow==11.0.0
pip-tools==7.4.1
propcache==0.2.1
pycparser==2.22
PyJWT==2.10.1
pyproject_hooks==1.2.0
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==75.8.0
sqlparse==0.5.2
twilio==9.4.3
urllib3==2.2.3
wheel==0.45.1
whitenoise==6.8.2
yarl==1.18.3
